env.concretize -> EnvironmentConcretizer(env).concretize()
