+++ lib/spack/spack/cmd/concretize.py
@@ -8,6 +8,7 @@ import spack.cmd
 import spack.cmd.common.arguments
 import spack.environment as ev
 import spack.llnl.util.tty as tty
+from spack.concretize import EnvironmentConcretizer
 from spack.llnl.string import plural
 
 description = "concretize an environment and write a lockfile"
@@ -41,7 +42,7 @@ def concretize(parser, args):
         tests = False
 
     with env.write_transaction():
-        concretized_specs = env.concretize(tests=tests)
+        concretized_specs = EnvironmentConcretizer(env).concretize(tests=tests)
         if not args.quiet:
             if concretized_specs:
                 tty.msg(f"Concretized {plural(len(concretized_specs), 'spec')}:")
