+++ lib/spack/spack/test/ci.py
@@ -17,6 +17,7 @@ import spack.llnl.util.filesystem as fs
 import spack.paths
 import spack.repo as repo
 import spack.util.git
+from spack.concretize import EnvironmentConcretizer
 from spack.spec import Spec
 from spack.test.conftest import MockHTTPResponse, RepoBuilder
 from spack.version import Version
@@ -360,7 +361,7 @@ def test_get_spec_filter_list(mutable_mock_env_path, mutable_mock_repo):
     e1 = ev.create("test")
     e1.add("mpileaks")
     e1.add("hypre")
-    e1.concretize()
+    EnvironmentConcretizer(e1).concretize()
 
     touched = {"libdwarf"}
 
@@ -418,7 +419,7 @@ def test_affected_specs_on_first_concretization(mutable_mock_env_path):
     e = ev.create("first_concretization")
     e.add("mpileaks~shared")
     e.add("mpileaks+shared")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
 
     affected_specs = ci.get_spec_filter_list(e, {"callpath"})
     mpileaks_specs = [s for s in affected_specs if s.name == "mpileaks"]
