+++ lib/spack/spack/test/cmd/ci.py
@@ -30,6 +30,7 @@ from spack.ci import gitlab as gitlab_generator
 from spack.ci.common import PipelineDag, PipelineOptions, SpackCIConfig
 from spack.ci.generator_registry import generator
 from spack.cmd.ci import FAILED_CREATE_BUILDCACHE_CODE
+from spack.concretize import EnvironmentConcretizer
 from spack.error import SpackError
 from spack.llnl.util.filesystem import mkdirp, working_dir
 from spack.schema.database_index import schema as db_idx_schema
@@ -539,7 +540,7 @@ spack:
         )
 
     with ev.Environment(env_dir) as env:
-        env.concretize()
+        EnvironmentConcretizer(env).concretize()
         env.write()
 
     shutil.copy(env_dir / "spack.yaml", tmp_path / "spack.yaml")
@@ -740,7 +741,7 @@ spack:
     with working_dir(tmp_path):
         env_cmd("create", "test", "./spack.yaml")
         with ev.read("test") as env:
-            env.concretize()
+            EnvironmentConcretizer(env).concretize()
 
             # Create environment variables as gitlab would do it
             os.environ.update(
@@ -812,7 +813,7 @@ spack:
             )
         env_cmd("create", "test", "./spack.yaml")
         with ev.read("test") as current_env:
-            current_env.concretize()
+            EnvironmentConcretizer(current_env).concretize()
             install_cmd("--keep-stage")
 
             concrete_spec = list(current_env.roots())[0]
@@ -1152,7 +1153,7 @@ spack:
     #          -> mpich
     env_hashes = {}
     with ev.read("test") as active_env:
-        active_env.concretize()
+        EnvironmentConcretizer(active_env).concretize()
         for s in active_env.all_specs():
             env_hashes[s.name] = s.dag_hash()
 
@@ -1369,7 +1370,7 @@ spack:
         )
 
     with working_dir(tmp_path), ev.Environment(".") as env:
-        env.concretize()
+        EnvironmentConcretizer(env).concretize()
         env.write()
 
     def fake_download_and_extract_artifacts(url, work_dir, merge_commit_test=True):
@@ -1885,7 +1886,7 @@ spack:
         manifest_data = json.load(fd)
 
     with ev.read("test") as active_env:
-        active_env.concretize()
+        EnvironmentConcretizer(active_env).concretize()
         for s in active_env.all_specs():
             assert s.dag_hash() in manifest_data
