+++ lib/spack/spack/test/cmd/deconcretize.py
@@ -5,6 +5,7 @@
 import pytest
 
 import spack.environment as ev
+from spack.concretize import EnvironmentConcretizer
 from spack.main import SpackCommand, SpackCommandError
 
 deconcretize = SpackCommand("deconcretize")
@@ -16,7 +17,7 @@ def test_env(mutable_mock_env_path, mock_packages):
     with ev.read("test") as e:
         e.add("pkg-a@2.0 foobar=bar ^pkg-b@1.0")
         e.add("pkg-a@1.0 foobar=bar ^pkg-b@0.9")
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
         e.write()
