+++ lib/spack/spack/test/cmd/env.py
@@ -37,6 +37,7 @@ import spack.util.environment
 import spack.util.spack_json as sjson
 import spack.util.spack_yaml
 from spack.cmd.env import _env_create
+from spack.concretize import EnvironmentConcretizer
 from spack.installer import PackageInstaller
 from spack.llnl.util.filesystem import readlink
 from spack.llnl.util.lang import dedupe
@@ -78,14 +79,14 @@ def setup_combined_multiple_env():
     test1 = ev.read("test1")
     with test1:
         add("mpich@1.0")
-        test1.concretize()
+        EnvironmentConcretizer(test1).concretize()
         test1.write()
 
     env("create", "test2")
     test2 = ev.read("test2")
     with test2:
         add("libelf")
-        test2.concretize()
+        EnvironmentConcretizer(test2).concretize()
         test2.write()
 
     env("create", "--include-concrete", "test1", "--include-concrete", "test2", "combined_env")
@@ -219,7 +220,7 @@ def installed_environment(
         with fs.working_dir(tmp_path):
             env("create", "test", "./spack.yaml")
             with ev.read("test") as current_environment:
-                current_environment.concretize()
+                EnvironmentConcretizer(current_environment).concretize()
                 current_environment.install_all(fake=True)
                 current_environment.write(regenerate=True)
 
@@ -294,7 +295,7 @@ def test_env_add_virtual():
     env("create", "test")
     e = ev.read("test")
     e.add("mpi")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
 
     assert len(e.concretized_roots) == 1
     spec = e.specs_by_hash[e.concretized_roots[0].hash]
@@ -473,7 +474,7 @@ def test_env_rename_independent(tmp_path: pathlib.Path):
 def test_concretize():
     e = ev.create("test")
     e.add("mpileaks")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
 
     assert len(e.concretized_roots) == 1
     assert e.concretized_roots[0].root == Spec("mpileaks")
@@ -482,7 +483,7 @@ def test_concretize():
 def test_env_specs_partition(install_mockery, mock_fetch):
     e = ev.create("test")
     e.add("cmake-client")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
 
     # Single not installed root spec.
     roots_already_installed, roots_to_install = e._partition_roots_by_install_status()
@@ -499,7 +500,7 @@ def test_env_specs_partition(install_mockery, mock_fetch):
 
     # One installed root, one not installed root.
     e.add("mpileaks")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
     roots_already_installed, roots_to_install = e._partition_roots_by_install_status()
     assert len(roots_already_installed) == 1
     assert len(roots_to_install) == 1
@@ -510,7 +511,7 @@ def test_env_specs_partition(install_mockery, mock_fetch):
 def test_env_install_all(install_mockery, mock_fetch):
     e = ev.create("test")
     e.add("cmake-client")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
     e.install_all(fake=True)
     spec = next(x for x in e.all_specs_generator() if x.name == "cmake-client")
     assert spec.installed
@@ -547,7 +548,7 @@ def test_env_install_include_concrete_env(
         mutable_config.set("concretizer:unify", unify)
         mutable_config.set("concretizer:reuse", reuse)
         combined.add("mpileaks")
-        combined.concretize()
+        EnvironmentConcretizer(combined).concretize()
         combined.write()
         install("--fake")
 
@@ -585,7 +586,7 @@ def test_env_roots_marked_explicit(install_mockery, mock_fetch):
     with ev.read("test") as e:
         # make implicit install a root of the env
         e.add(dependency[0].name)
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
         e.install_all()
 
     explicit = spack.store.STORE.db.query(explicit=True)
@@ -638,7 +639,7 @@ def test_env_definition_symlink(install_mockery, mock_fetch, tmp_path: pathlib.P
     os.symlink(filepath, filepath_mid)
     os.symlink(filepath_mid, e.manifest_path)
 
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
     e.write()
 
     assert os.path.islink(e.manifest_path)
@@ -686,10 +687,10 @@ def test_remove_after_concretize():
     e = ev.create("test")
 
     e.add("mpileaks")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
 
     e.add("python")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
 
     e.remove("mpileaks")
     assert Spec("mpileaks") not in e.user_specs
@@ -708,12 +709,12 @@ def test_remove_before_concretize(mutable_config):
     with ev.create("test") as e:
         mutable_config.set("concretizer:unify", True)
         e.add("mpileaks")
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
         assert len(e.concretized_roots) == 1
         assert e.concrete_roots()[0].satisfies("mpileaks")
 
         e.remove("mpileaks")
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
         assert not e.concretized_roots
 
 
@@ -800,7 +801,7 @@ def test_bad_remove_included_env():
     with test:
         add("mpileaks")
 
-    test.concretize()
+    EnvironmentConcretizer(test).concretize()
     test.write()
 
     env("create", "--include-concrete", "test", "combined_env")
@@ -816,7 +817,7 @@ def test_force_remove_included_env():
     with test:
         add("mpileaks")
 
-    test.concretize()
+    EnvironmentConcretizer(test).concretize()
     test.write()
 
     env("create", "--include-concrete", "test", "combined_env")
@@ -891,7 +892,7 @@ def test_env_activate_broken_view(
 def test_to_lockfile_dict():
     e = ev.create("test")
     e.add("mpileaks")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
     context_dict = e._to_lockfile_dict()
 
     e_copy = ev.create("test_copy")
@@ -924,7 +925,7 @@ spack:
   - libelf
 """
     )
-    before.concretize()
+    EnvironmentConcretizer(before).concretize()
     before.write()
 
     # user modifies yaml externally to spack and removes hypre
@@ -939,7 +940,7 @@ spack:
         )
 
     after = ev.read("test")
-    after.concretize()
+    EnvironmentConcretizer(after).concretize()
     after.write()
 
     read = ev.read("test")
@@ -966,7 +967,7 @@ spack:
 """
     )
     with e1:
-        e1.concretize()
+        EnvironmentConcretizer(e1).concretize()
         e1.write()
 
     # By reading into a second environment, we force a round trip to json
@@ -990,7 +991,7 @@ spack:
   - libelf
 """
     )
-    e1.concretize()
+    EnvironmentConcretizer(e1).concretize()
     e1.write()
 
     e2 = _env_create("test2", init_file=e1.lock_path)
@@ -1015,7 +1016,7 @@ spack:
   - libelf
 """
     )
-    e1.concretize()
+    EnvironmentConcretizer(e1).concretize()
     e1.write()
 
     e2 = _env_create("test2", init_file=e1.manifest_path)
@@ -1050,7 +1051,7 @@ spack:
         spack.config.set("develop", dev_config)
         fs.touch(os.path.join(e1.path, "libelf"))
 
-    e1.concretize()
+    EnvironmentConcretizer(e1).concretize()
     e1.write()
 
     e2 = _env_create("test2", init_file="test" if use_name else e1.path)
@@ -1167,7 +1168,7 @@ packages:
     test_scope = spack.config.InternalConfigScope("env-external-test", data=external_config_dict)
     with spack.config.override(test_scope):
         e = ev.create("test", manifest_file)
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
         # Note: normally installing specs in a test environment requires doing
         # a fake install, but not for external specs since no actions are
         # taken to install them. The installation commands also include
@@ -1223,7 +1224,7 @@ spack:
 """
     )
     with e:
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
 
     mpileaks_hash = next(x.hash for x in e.concretized_roots if x.root == Spec("mpileaks"))
     mpileaks = e.specs_by_hash[mpileaks_hash]
@@ -1259,7 +1260,7 @@ spack:
     )
 
     with ev.Environment(env_root) as e:
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
 
     # We've created an environment with included config file (which does
     # exist). Now we remove it and check that we get a sensible error.
@@ -1309,7 +1310,7 @@ spack:
     )
 
     with e:
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
 
     mpileaks_hash = next(x.hash for x in e.concretized_roots if x.root == Spec("mpileaks"))
     mpileaks = e.specs_by_hash[mpileaks_hash]
@@ -1372,7 +1373,7 @@ spack:
 
     e = ev.Environment(env_root)
     with e:
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
 
     mpileaks_hash = next(x.hash for x in e.concretized_roots if x.root == Spec("mpileaks"))
     mpileaks = e.specs_by_hash[mpileaks_hash]
@@ -1538,7 +1539,7 @@ def test_env_with_included_config_scope(mutable_mock_env_path, packages_file):
     # packages.yaml file.
     e = ev.Environment(env_root)
     with e:
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
 
     mpileaks_hash = next(x.hash for x in e.concretized_roots if x.root == Spec("mpileaks"))
     mpileaks = e.specs_by_hash[mpileaks_hash]
@@ -1561,7 +1562,7 @@ def test_env_with_included_config_var_path(tmp_path: pathlib.Path, packages_file
 
     e = ev.Environment(env_path)
     with e:
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
 
     mpileaks_hash = next(x.hash for x in e.concretized_roots if x.root == Spec("mpileaks"))
     mpileaks = e.specs_by_hash[mpileaks_hash]
@@ -1601,7 +1602,7 @@ spack:
 
     e = ev.Environment(tmp_path)
     with e:
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
 
     mpileaks_hash = next(x.hash for x in e.concretized_roots if x.root == Spec("mpileaks"))
     mpileaks = e.specs_by_hash[mpileaks_hash]
@@ -1652,7 +1653,7 @@ packages:
 
     e = ev.Environment(tmp_path)
     with e:
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
 
     mpileaks_hash = next(x.hash for x in e.concretized_roots if x.root == Spec("mpileaks"))
     mpileaks = e.specs_by_hash[mpileaks_hash]
@@ -1907,7 +1908,7 @@ def test_indirect_build_dep(repo_builder: RepoBuilder):
         _env_create("test", with_view=False)
         e = ev.read("test")
         e.add(x_spec)
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
         e.write()
 
         e_read = ev.read("test")
@@ -1949,7 +1950,7 @@ def test_store_different_build_deps(repo_builder: RepoBuilder):
         e = ev.read("test")
         e.add(y_spec)
         e.add(x_spec)
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
         e.write()
 
         e_read = ev.read("test")
@@ -2055,7 +2056,7 @@ def test_env_include_concrete_env_yaml(env_name):
 
     with test:
         add("mpileaks")
-    test.concretize()
+    EnvironmentConcretizer(test).concretize()
     test.write()
 
     environ = "test" if env_name else test.path
@@ -2136,7 +2137,7 @@ def test_env_include_concrete_add_env():
     with new_env:
         add("mpileaks")
 
-    new_env.concretize()
+    EnvironmentConcretizer(new_env).concretize()
     new_env.write()
 
     # add new env to combined
@@ -2149,7 +2150,7 @@ def test_env_include_concrete_add_env():
     assert new_env.path not in lockfile_as_dict["include_concrete"].keys()
 
     # concretize combined env with new env
-    combined.concretize()
+    EnvironmentConcretizer(combined).concretize()
     combined.write()
 
     # assert changes
@@ -2172,7 +2173,7 @@ def test_env_include_concrete_remove_env():
     assert test2.path in lockfile_as_dict["include_concrete"].keys()
 
     # reconcretize combined
-    combined.concretize()
+    EnvironmentConcretizer(combined).concretize()
     combined.write()
 
     # assert test2 is not in combined's lockfile
@@ -2198,7 +2199,7 @@ def configure_reuse(reuse_mode, combined_env) -> Optional[ev.Environment]:
         override_env = ev.read("external_test")
         with override_env:
             add("mpich@1.0 +debug")
-        override_env.concretize()
+        EnvironmentConcretizer(override_env).concretize()
         override_env.write()
 
         # Reuse from the environment that is not included.
@@ -2255,7 +2256,7 @@ def test_env_include_concrete_reuse(reuse_mode):
         # Add mpileaks to the combined environment
         with combined:
             add("mpileaks")
-            combined.concretize()
+            EnvironmentConcretizer(combined).concretize()
         comb_specs_by_hash = combined.specs_by_hash
 
         # create reference env with mpileaks that does not use reuse
@@ -2264,7 +2265,7 @@ def test_env_include_concrete_reuse(reuse_mode):
         ref_env = ev.read("new")
         with ref_env:
             add("mpileaks")
-        ref_env.concretize()
+        EnvironmentConcretizer(ref_env).concretize()
         ref_specs_by_hash = ref_env.specs_by_hash
 
         # Ensure that the mpich used by the mpileaks is the mpich from the reused test environment
@@ -2299,7 +2300,7 @@ def test_env_include_concrete_env_reconcretized(mutable_config, unify):
 
     with combined:
         mutable_config.set("concretizer:unify", unify)
-        combined.concretize()
+        EnvironmentConcretizer(combined).concretize()
         combined.write()
 
     with open(combined.lock_path, encoding="utf-8") as f:
@@ -2318,7 +2319,7 @@ def test_concretize_include_concrete_env():
     # Update test1 environment
     with test1:
         add("mpileaks")
-    test1.concretize()
+    EnvironmentConcretizer(test1).concretize()
     test1.write()
 
     # Check the test1 environment includes mpileaks, while the combined environment does not
@@ -2326,7 +2327,7 @@ def test_concretize_include_concrete_env():
     assert Spec("mpileaks") not in combined.included_concretized_user_specs[test1.path]
 
     # If we update the combined environment, it will include mpileaks too
-    combined.concretize()
+    EnvironmentConcretizer(combined).concretize()
     combined.write()
     assert Spec("mpileaks") in combined.included_concretized_user_specs[test1.path]
 
@@ -2336,14 +2337,14 @@ def test_concretize_nested_include_concrete_envs():
     test1 = ev.read("test1")
     with test1:
         add("zlib")
-    test1.concretize()
+    EnvironmentConcretizer(test1).concretize()
     test1.write()
 
     env("create", "--include-concrete", "test1", "test2")
     test2 = ev.read("test2")
     with test2:
         add("libelf")
-    test2.concretize()
+    EnvironmentConcretizer(test2).concretize()
     test2.write()
 
     env("create", "--include-concrete", "test2", "test3")
@@ -2365,7 +2366,7 @@ def test_concretize_nested_included_concrete():
     test1 = ev.read("test1")
     with test1:
         add("zlib")
-    test1.concretize()
+    EnvironmentConcretizer(test1).concretize()
     test1.write()
 
     # test2 should include test1 with zlib
@@ -2373,7 +2374,7 @@ def test_concretize_nested_included_concrete():
     test2 = ev.read("test2")
     with test2:
         add("libelf")
-    test2.concretize()
+    EnvironmentConcretizer(test2).concretize()
     test2.write()
 
     assert Spec("zlib") in test2.included_concretized_user_specs[test1.path]
@@ -2382,7 +2383,7 @@ def test_concretize_nested_included_concrete():
     with test1:
         remove("zlib")
         add("mpileaks")
-    test1.concretize()
+    EnvironmentConcretizer(test1).concretize()
     test1.write()
 
     # test3 should include the latest concretization of test1
@@ -2390,7 +2391,7 @@ def test_concretize_nested_included_concrete():
     test3 = ev.read("test3")
     with test3:
         add("callpath")
-    test3.concretize()
+    EnvironmentConcretizer(test3).concretize()
     test3.write()
 
     included_specs = test3.included_concretized_user_specs[test1.path]
@@ -2423,9 +2424,9 @@ def test_concretize_nested_included_concrete():
     # to remove zlib and write it out so it can be picked up by test4.
     # Re-concretize test4 to reflect the re-concretization of included test2
     # and ensure that its included specs are up-to-date
-    test2.concretize()
+    EnvironmentConcretizer(test2).concretize()
     test2.write()
-    test4.concretize()
+    EnvironmentConcretizer(test4).concretize()
 
     concrete_specs = [s for s, _ in test4.concretized_specs()]
     assert Spec("zlib") not in concrete_specs
@@ -2771,7 +2772,7 @@ spack:
     with fs.working_dir(str(tmp_path)):
         env("create", "test", "./spack.yaml")
         with ev.read("test") as e:
-            e.concretize()
+            EnvironmentConcretizer(e).concretize()
 
             before_user = e.user_specs.specs
             concretized_roots_before = e.concretized_roots
@@ -3361,7 +3362,7 @@ def test_concretize_user_specs_together(mutable_config):
         # Concretize a first time using 'mpich' as the MPI provider
         e.add("mpileaks")
         e.add("mpich")
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
 
         assert all("mpich" in spec for _, spec in e.concretized_specs())
         assert all("mpich2" not in spec for _, spec in e.concretized_specs())
@@ -3374,15 +3375,15 @@ def test_concretize_user_specs_together(mutable_config):
 
         # Concretizing without invalidating the concrete spec for mpileaks fails
         with pytest.raises(exc_cls):
-            e.concretize()
-        e.concretize(force=True)
+            EnvironmentConcretizer(e).concretize()
+        EnvironmentConcretizer(e).concretize(force=True)
 
         assert all("mpich2" in spec for _, spec in e.concretized_specs())
         assert all("mpich" not in spec for _, spec in e.concretized_specs())
 
         # Concretize again without changing anything, check everything
         # stays the same
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
 
         assert all("mpich2" in spec for _, spec in e.concretized_specs())
         assert all("mpich" not in spec for _, spec in e.concretized_specs())
@@ -3399,7 +3400,7 @@ def test_duplicate_packages_raise_when_concretizing_together(mutable_config):
         match = r"You could consider setting `concretizer:unify`"
 
         with pytest.raises(exc_cls, match=match):
-            e.concretize()
+            EnvironmentConcretizer(e).concretize()
 
 
 def test_env_write_only_non_default():
@@ -3437,7 +3438,7 @@ spack:
 
     # concretize
     with ev.read("test") as e:
-        concretize()
+        EnvironmentConcretizer(e).concretize()
         e.write()
 
         with open(e.manifest_path, "r", encoding="utf-8") as f:
@@ -3491,12 +3492,12 @@ spack:
 
     # Concretize a first time and create a lockfile
     with ev.Environment(str(tmp_path)) as e:
-        concretize()
+        EnvironmentConcretizer(e).concretize()
         libelf_first_hash = extract_dag_hash(e)
 
     # Check that a second run won't error
     with ev.Environment(str(tmp_path)) as e:
-        concretize()
+        EnvironmentConcretizer(e).concretize()
         libelf_second_hash = extract_dag_hash(e)
 
     assert libelf_first_hash == libelf_second_hash
@@ -3532,7 +3533,7 @@ spack:
         spack.environment.environment.EnvironmentManifestFile, "flush", _write_helper_raise
     )
     with ev.Environment(str(tmp_path)) as e:
-        e.concretize(force=True)
+        EnvironmentConcretizer(e).concretize(force=True)
         with pytest.raises(RuntimeError):
             e.clear()
             e.write()
@@ -3607,7 +3608,7 @@ def test_custom_version_concretize_together(mutable_config):
         # Concretize a first time using 'mpich' as the MPI provider
         e.add("hdf5@=myversion")
         e.add("mpich")
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
         assert any(spec.satisfies("hdf5@myversion") for _, spec in e.concretized_specs())
 
 
@@ -3729,7 +3730,7 @@ def test_virtual_spec_concretize_together(mutable_config):
     with ev.create("virtual_spec") as e:
         mutable_config.set("concretizer:unify", True)
         e.add("mpi")
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
         assert any(s.package.provides("mpi") for _, s in e.concretized_specs())
 
 
@@ -4366,7 +4367,7 @@ def test_unify_when_possible_works_around_conflicts(mutable_config):
         e.add("mpileaks+opt")
         e.add("mpileaks~opt")
         e.add("mpich")
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
 
         assert len([x for x in e.all_specs() if x.satisfies("mpileaks")]) == 2
         assert len([x for x in e.all_specs() if x.satisfies("mpileaks+opt")]) == 1
@@ -4466,7 +4467,7 @@ spack:
     )
     env("create", "disabled", str(spack_yaml))
     with ev.read("disabled") as e:
-        e.concretize()
+        EnvironmentConcretizer(e).concretize()
 
     assert len(e.views) == 0
     assert not os.path.exists(e.view_path_default)
@@ -4672,7 +4673,7 @@ def test_concretized_specs_and_include_concrete(mutable_config):
         mutable_config.set(
             "packages", {"mpileaks": {"require": ["@2.2"]}, "libelf": {"require": ["@0.8.12"]}}
         )
-        included_pairs = e.concretize()
+        included_pairs = EnvironmentConcretizer(e).concretize()
         e.write()
 
     env("create", "--include-concrete", "included-env", "main-env")
@@ -4681,7 +4682,7 @@ def test_concretized_specs_and_include_concrete(mutable_config):
         e.add("libelf@0.8.12")
         e.add("pkg-a")
         mutable_config.set("packages", {"mpileaks": {"require": ["@2.3"]}})
-        spec_pairs = e.concretize()
+        spec_pairs = EnvironmentConcretizer(e).concretize()
         concretized_specs = list(e.concretized_specs())
         assert list(dedupe(spec_pairs + included_pairs)) == concretized_specs
         assert len(concretized_specs) == 5
