+++ lib/spack/spack/test/cmd/stage.py
@@ -11,6 +11,7 @@ import spack.environment as ev
 import spack.package_base
 import spack.traverse
 from spack.cmd.stage import StageFilter
+from spack.concretize import EnvironmentConcretizer
 from spack.main import SpackCommand, SpackCommandError
 from spack.spec import Spec
 from spack.version import Version
@@ -72,7 +73,7 @@ def test_stage_with_env_outside_env(mutable_mock_env_path, monkeypatch):
 
     e = ev.create("test")
     e.add("mpileaks")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
 
     with e:
         stage("trivial-install-test-package")
@@ -90,7 +91,7 @@ def test_stage_with_env_inside_env(mutable_mock_env_path, monkeypatch):
 
     e = ev.create("test")
     e.add("mpileaks@=100.100")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
 
     with e:
         stage("mpileaks")
@@ -102,7 +103,7 @@ def test_stage_full_env(mutable_mock_env_path, monkeypatch):
 
     e = ev.create("test")
     e.add("mpileaks@=100.100")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
 
     # list all the package names that should be staged
     expected, externals = set(), set()
@@ -151,7 +152,7 @@ def test_stage_spec_filters(
 ):
     e = ev.create("test")
     e.add("mpileaks@=100.100")
-    e.concretize()
+    EnvironmentConcretizer(e).concretize()
     all_specs = e.all_specs()
 
     def is_installed(self):
