+++ lib/spack/spack/test/concretization/core.py
@@ -41,6 +41,7 @@ import spack.util.file_cache
 import spack.util.hash
 import spack.util.spack_yaml as syaml
 import spack.variant as vt
+from spack.concretize import EnvironmentConcretizer
 from spack.externals import ExternalDependencyError
 from spack.installer import PackageInstaller
 from spack.solver.asp import Result
@@ -585,7 +586,7 @@ spack:
         )
 
         with ev.Environment(tmp_path) as e:
-            e.concretize()
+            EnvironmentConcretizer(e).concretize()
             for root in e.roots():
                 if root.satisfies("%gcc"):
                     assert root["dt-diamond-left"].satisfies("%gcc")
